fix: resolve all PHPStan Level 9 errors to unblock push

+++ apps/backend/app/Actions/Finance/Transaction/GetTransactionSummaryAction.php
@@ -28,7 +28,7 @@ class GetTransactionSummaryAction extends BaseAction
      */
     public function execute(int $userId, ?int $month, ?int $year, int $budgetCycleStart): array
     {
-        $version = Cache::get("transaction_summary_version_{$userId}", 1);
+        $version = (int) Cache::get("transaction_summary_version_{$userId}", 1);
         $cacheKey = "transaction_summary_{$userId}_".($month ?? 'all').'_'.($year ?? 'all')."_{$budgetCycleStart}_v{$version}";
 
         return Cache::remember($cacheKey, 600, function () use ($userId, $month, $year, $budgetCycleStart) {
